swarmshield/requirements.txt
crewai==0.1.0
crewai-tools==0.1.0
python-dotenv==1.0.0
pydantic==2.0.0
pyyaml==6.0
networkx==3.2
scikit-learn==1.3.0
tensorflow==2.13.0
torch==2.0.0
torch-geometric==2.3.0
scapy==2.5.0
pyshark==0.6.0
deap==1.4.1
streamlit==1.28.0
pandas==2.0.0
numpy==1.24.0
matplotlib==3.7.0
seaborn==0.12.0
plotly==5.17.0
requests==2.31.0
pytest==7.4.0
pytest-cov==4.1.0
black==23.9.0
flake8==6.1.0
mypy==1.5.0
ipython==8.14.0
jupyter==1.0.0
notebook==7.0.0
